requirements.txt
flask 
google-genai 
